requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
flask
flask_sqlalchemy
flask_cors
python-dotenv
PyJWT
werkzeug
bleach
alembic
psycopg2-binary
requests>=2.31.0
pgvector>=0.2.5

# ── Testing ───────────────────────────────────────────────────────────────────
pytest

# ── Document Parsing (Teacher RAG) ───────────────────────────────────────────
pypdf>=4.2
python-docx>=1.1
boto3>=1.34

# ── AI / Machine Learning ─────────────────────────────────────────────────────
# Core deep learning framework
# TensorFlow wheels are not available for Python 3.14 yet.
tensorflow>=2.13,<2.17; python_version < "3.13"

# Image processing
Pillow>=9.0

# Scientific computing (likely already present via tensorflow, listed explicitly)
numpy>=1.24
scipy>=1.10
scikit-image>=0.22
scikit-learn>=1.3             # class_weight, metrics, confusion_matrix

# Visualisation (training curves, confusion matrix)
matplotlib>=3.7
seaborn>=0.12

# Explainability
shap>=0.44.0

# TensorFlow.js export  — converts .h5 → browser-ready model.json + weights.bin
# Optional for training/export workflows only.
tensorflowjs>=4.0; python_version < "3.13"
